linux: relocate PIDFD definitions to src/new
