+++ src/new/linux_uapi/linux/mod.rs
@@ -6,3 +6,4 @@ pub(crate) mod can;
 pub(crate) mod keyctl;
 pub(crate) mod membarrier;
 pub(crate) mod netlink;
+pub(crate) mod pidfd;
